requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas==1.1.5
azure.storage.blob==12.16.0
joblib==1.2.0
sklearn==0.0
psycopg2-binary
cryptography
datetime
scrypt==0.8.20
PyJWT==2.1.0